pyproject.toml
[project]
name = "ai-dlc-prompts-stage0"
version = "0.1.0"
requires-python = ">=3.12"

dependencies = [
  "jinja2>=3.1",
  "typer>=0.12",
  "jsonschema>=4.21",
  "rich>=14.0.0",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "ruff>=0.4",
]

[tool.uv]          # optional – tweak resolver settings here

[project.scripts]
ai-dlc = "cli.main:app"

